Promote JvmModifier classes to top level
